solved issues with module sample_gender, combined files and renamed  them according to common practice
