--- multiqc_cmgg/modules/ngsbits/__init__.py
@@ -1,3 +1,0 @@
-from .ngsbits import MultiqcModule
-
-__all__ = ["MultiqcModule"]
