+++ multiqc_cmgg/modules/sample_gender/__init__.py
@@ -1,0 +1,3 @@
+from .sample_gender import MultiqcModule
+
+__all__ = ["MultiqcModule"]
