+++ test_data/ngsbits/test1_xy.tsv
@@ -1,0 +1,2 @@
+#file	gender	reads_chry	reads_chrx	ratio_chry_chrx
+HG00096.mapped.ILLUMINA.bwa.GBR.low_coverage.20120522.bam.cram	male	819372	3670740	0.2232
