+++ test_data/ngsbits/test2_xy.tsv
@@ -1,0 +1,2 @@
+#file	gender	reads_chry	reads_chrx	ratio_chry_chrx
+HG00097.mapped.ILLUMINA.bwa.GBR.low_coverage.20130415.bam.cram	female	56364	14308554	0.0039
