+++ test_data/ngsbits/test3_xy.tsv
@@ -1,0 +1,2 @@
+#file	gender	reads_chry	reads_chrx	ratio_chry_chrx
+HG00099.mapped.ILLUMINA.bwa.GBR.low_coverage.20130415.bam.cram	female	44853	11716394	0.0038
