+++ test_data/ngsbits/test_xy.tsv
@@ -1,0 +1,4 @@
+#file	gender	reads_chry	reads_chrx	ratio_chry_chrx
+HG00096.mapped.ILLUMINA.bwa.GBR.low_coverage.20120522.bam.cram	male	819372	3670740	0.2232
+HG00097.mapped.ILLUMINA.bwa.GBR.low_coverage.20130415.bam.cram	female	56364	14308554	0.0039
+HG00099.mapped.ILLUMINA.bwa.GBR.low_coverage.20130415.bam.cram	female	44853	11716394	0.0038
